backend/agents/requirements.txt
google-genai
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
redis
sqlalchemy
psycopg2-binary
pytest
pytest-asyncio
pillow
websockets
